# cmd/server/docs/swagger.yaml
basePath: /
definitions:
  github_com_GiorgiUbiria_banking_system_internal_models.Account:
    properties:
      balance:
        type: number
      createdAt:
        type: string
      currency:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
  github_com_GiorgiUbiria_banking_system_internal_models.Transaction:
    properties:
      createdAt:
        type: string
      currency:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      status:
        description: pending | completed | failed
        type: string
      type:
        description: transfer | exchange
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_handlers.LoginResponse:
    properties:
      token:
        type: string
    type: object
  internal_handlers.TransferRequest:
    properties:
      amount:
        type: string
      from_account_id:
        type: integer
      to_account_id:
        type: integer
    type: object
  internal_handlers.TransferResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simplified banking platform for the assessment
  title: Mini Banking API
  version: "1.0"
paths:
  /accounts:
    get:
      description: Get all accounts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_GiorgiUbiria_banking_system_internal_models.Account'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: List user accounts
      tags:
      - accounts
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.LoginResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "401":
          description: invalid email or password
          schema:
            type: string
      summary: User login
      tags:
      - auth
  /transactions:
    get:
      description: Get all transactions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_GiorgiUbiria_banking_system_internal_models.Transaction'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: List user transactions
      tags:
      - transactions
  /transactions/transfer:
    post:
      consumes:
      - application/json
      description: Transfer money between accounts for the authenticated user
      parameters:
      - description: Transfer request
        in: body
        name: transfer
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.TransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.TransferResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Transfer money between accounts
      tags:
      - transactions
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"